move tables to one file and delete oldTables
